+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `isExported` util function to avoid importing `ast` package
+
 
 ## [0.3.0] - 2017-12-21